+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/actions/TyeHostShutdownAction.kt
@@ -10,28 +10,28 @@ import com.intellij.openapi.progress.Task
 import com.intellij.openapi.project.Project
 import kotlinx.coroutines.runBlocking
 
-class RefreshDashboardAction : AnAction() {
+class TyeHostShutdownAction : AnAction() {
 
     override fun actionPerformed(e: AnActionEvent) {
-        val task = object : Task.Backgroundable(e.project, "Refresh dashboard") {
+        val task = object : Task.Backgroundable(e.project, "Tye shutdown") {
             override fun run(indicator: ProgressIndicator) {
                 if (indicator.isCanceled) {
                     return
                 }
 
                 indicator.isIndeterminate = true
-                indicator.text = "Refreshing tye dashboard"
+                indicator.text = "Tye host shutdown"
 
-                updateDashboard(project!!)
+                shutdown(project!!)
             }
         }
         ProgressManager.getInstance().run(task)
     }
 
-    private fun updateDashboard(project: Project) {
+    private fun shutdown(project: Project) {
         val tyeApiService = project.service<TyeApplicationService>()
         runBlocking {
-            tyeApiService.update()
+            tyeApiService.shutdown()
         }
     }
 }