+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/api/TyeApiClient.kt
@@ -5,6 +5,7 @@ import io.ktor.client.HttpClient
 import io.ktor.client.engine.cio.CIO
 import io.ktor.client.features.json.JsonFeature
 import io.ktor.client.features.json.serializer.KotlinxSerializer
+import io.ktor.client.request.delete
 import io.ktor.client.request.get
 
 class TyeApiClient {
@@ -14,7 +15,7 @@ class TyeApiClient {
         }
     }
 
-    suspend fun getServices(host: String): List<ServiceDto> {
-        return httpClient.get("$host/api/v1/services")
-    }
+    suspend fun getServices(host: String): List<ServiceDto> = httpClient.get("$host/api/v1/services")
+
+    suspend fun controlPlaneShutdown(host: String) = httpClient.delete<Unit>("$host/api/v1/control")
 }